refactor(skills): rename account skill to vesta-cloud-account
Align the skill identity with its CLI command (already
`vesta-cloud-account`) and disambiguate from the billing/email "account"
senses scattered across the codebase.

- skill dir account/ -> vesta-cloud-account/ and name: in SKILL.md
- Python package account_cli -> vc_account_cli (pyproject script +
  wheel packages, test imports)
- default-skills.txt entry + regenerated index.json
- onboard references (SKILL.md, config/referral_store/test comments)
  point at the new name
- drop the obsolete 2026-06-account-default-skill migration (the
  stateless default-skill reconciler now owns installing missing
  defaults) and add 2026-07-rename-account-skill, which only clears the
  stale account sparse-cone entry on existing boxes
