+++ agent/core/default-skills.txt
@@ -1,4 +1,3 @@
-account
 birth
 browser
 claude-code
@@ -16,5 +15,6 @@ skills-registry
 tasks
 timezone
 upstream-pr
+vesta-cloud-account
 voice
 what-day
