+++ agent/skills/onboard/cli/src/onboard_cli/referral_store.py
@@ -1,8 +1,8 @@
-"""Shared on-disk referral code — the bridge from the `account` skill.
+"""Shared on-disk referral code — the bridge from the `vesta-cloud-account` skill.
 
-The account skill and this one are separate Python packages (they can't import
-each other), so this file mirrors `account_cli.referral_store` byte-for-byte on
-the path: the account skill is the only place the code is ever SET
+The vesta-cloud-account skill and this one are separate Python packages (they can't import
+each other), so this file mirrors `vc_account_cli.referral_store` byte-for-byte on
+the path: the vesta-cloud-account skill is the only place the code is ever SET
 (`vesta-cloud-account set-referral`); this reader is the only thing this skill
 does with it. Plain UTF-8 text = the code, stripped. A missing or empty file
 means "no code configured".
