+++ agent/skills/vesta-cloud-account/cli/tests/test_cli.py
@@ -6,9 +6,9 @@ import json
 
 import pytest
 
-from account_cli import cli as cli_mod
-from account_cli import referral_store
-from account_cli.client import AccountError
+from vc_account_cli import cli as cli_mod
+from vc_account_cli import referral_store
+from vc_account_cli.client import AccountError
 
 
 @pytest.fixture(autouse=True)
